security config (not yet)

+++ src/main/java/com/example/kiennt_demo2/repository/EmployeeRepository.java
@@ -9,5 +9,5 @@ import java.util.List;
 @Repository
 public interface EmployeeRepository extends JpaRepository<Employee, Integer> {
     List<Employee> findAll();
-    Employee findById(int id);
+    Employee findById(long id);
 }
